requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
flower==2.0.1
humanize==4.15.0
kombu==5.6.2
packaging==26.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
pytz==2025.2
redis==7.1.0
six==1.17.0
sqlparse==0.5.5
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.5.3
